Moved customer interfaces
