+++ client/src/interfaces/customers/customerSearch.ts
