+++ client/src/redux/thunks/searchCustomersThunks.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
 import appAxiosInstance from "../../api/appAxiosInstance";
-import { CustomersInList } from "../../interfaces/customers";
+import { CustomersInList } from "../../interfaces/customers/customerSearch";
 import { RootState } from "../store";
 
 export const fetchCustomersThunk = createAsyncThunk(
